removed translation map file and filter options type
